Maximize backwards compatible with python2
