+++ spec/python3/query_hook_spec.rb
@@ -11,4 +11,14 @@ describe Python3QueryHook do
     let(:request) { struct query: 'raise Exception, "bar"' }
     it { expect(result[1]).to eq :errored }
   end
+
+  context 'passes when standalone query is valid and returns a string.' do
+    let(:request) { struct query: '"foo"' }
+    it { expect(result).to eq ["'foo'\n", :passed] }
+  end
+
+  context 'passes when standalone query is valid and has utf8 chars.' do
+    let(:request) { struct query: '"fó" + "ò"' }
+    it { expect(result).to eq ["'fóò'\n", :passed] }
+  end
 end
